# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: User Management Control Panel API
  version: 0.1.0
  description: |
    Control-plane API for the user-management + licensing system.

    Covers admin authentication, organization/user/RBAC management, plan & subscription
    administration, license JWKS publication, key rotation, usage ingestion, audit log
    access, SSO configuration, API key management, and the domain event stream.

    Two authentication schemes are supported:
      - **bearerAuth**: short-lived admin session JWT, issued by `POST /api/v1/auth/login`.
      - **apiKeyAuth**: long-lived programmatic key, sent as `Authorization: ApiKey <token>`.

    The public `/.well-known/jwks.json` endpoint is unauthenticated.

    Error responses follow [RFC 7807 Problem Details](https://datatracker.ietf.org/doc/html/rfc7807).

    > **Caveat:** the license-issuance endpoint (`POST .../licenses`) referenced in the
    > implementation plan is *not yet wired up* in `control-panel-api`. It is documented
    > here as the intended contract for Phase 1 closure.
  contact:
    name: Control Panel Team
    email: support@control-panel.example.com
  license:
    name: Proprietary
servers:
  - url: https://control-panel.example.com
    description: Production
  - url: https://staging.control-panel.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local dev
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /api/v1/orgs:
    get:
      tags: [orgs]
      summary: List organizations the current user belongs to.
      responses:
        '200':
          description: Paged org list (single page in the current implementation).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationPage'
    post:
      tags: [orgs]
      summary: Create an organization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreate'
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/orgs/{orgId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [orgs]
      summary: Fetch an organization.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orgs/{orgId}/members:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    get:
      tags: [orgs]
      summary: List members of an organization.
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMemberPage'
    post:
      tags: [orgs]
      summary: Add a user as a member of the organization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgInvite'
      responses:
        '201':
          description: Member added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMember'
  /api/v1/orgs/{orgId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
      - $ref: '#/components/parameters/UserIdParam'
    delete:
      tags: [orgs]
      summary: Remove a member.
      responses:
        '204':
          description: Removed.
  /api/v1/orgs/{orgId}/transfer-owner:
    parameters:
      - $ref: '#/components/parameters/OrgIdParam'
    post:
      tags: [orgs]
      summary: Transfer the OWNER role to another member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [newOwnerUserId]
              properties:
                newOwnerUserId: { type: string, format: uuid }
      responses:
        '204':
          description: Owner transferred.

  # ============================================================
  # users
  # ============================================================
...
  /api/v1/subscriptions/{subscriptionId}/licenses:
    parameters:
      - $ref: '#/components/parameters/SubscriptionIdParam'
    get:
      tags: [licenses]
      summary: '[Planned] List licenses issued for a subscription.'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LicenseToken'
    post:
      tags: [licenses]
      summary: '[Planned] Issue a new license for the subscription.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LicenseIssueRequest'
      responses:
        '201':
          description: "License issued. Set `Accept: application/vnd.example.license+json` to receive only the `.lic` envelope for direct download."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseIssueResponse'
            application/vnd.example.license+json:
              schema:
                $ref: '#/components/schemas/LicenseEnvelope'
  /api/v1/licenses/{jti}:
    parameters:
      - in: path
        name: jti
        required: true
        schema: { type: string }
    get:
      tags: [licenses]
      summary: '[Planned] Get a license record by jti.'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseToken'
  /api/v1/licenses/{jti}/revoke:
    parameters:
      - in: path
        name: jti
        required: true
        schema: { type: string }
    post:
      tags: [licenses]
      summary: '[Planned] Revoke a license.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonRequest'
      responses:
        '200':
          description: Revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LicenseToken'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin session JWT from `POST /api/v1/auth/login`.
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Long-lived programmatic key.
        Format: `Authorization: ApiKey <token>` where `<token>` is the value returned
        once by `POST /api/v1/orgs/{orgId}/api-keys`.
  parameters:
    OrgIdParam:
      in: path
      name: orgId
      required: true
      schema: { type: string, format: uuid }
    UserIdParam:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
    SubscriptionIdParam:
      in: path
      name: subscriptionId
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request body or parameters invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    NotFound:
      description: Resource does not exist or is hidden from this caller.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflicting state (e.g. unique-constraint violation).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Bad Request' }
        status: { type: integer, example: 400 }
        detail: { type: string, example: 'Field `email` is required.' }
        instance: { type: string, format: uri }
        code: { type: string, example: 'validation_failed' }
      required: [title, status]
    ReasonRequest:
      type: object
      properties:
        reason: { type: string }
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
      required: [total, page, size]

    # ---- auth ----
...
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, slug, name]
    OrganizationCreate:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{1,62}[a-z0-9]$'
        name:
          type: string
          maxLength: 255
    OrganizationPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Organization' }
          required: [items]
    OrgMember:
      type: object
      properties:
        orgId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER, VIEWER]
        addedAt: { type: string, format: date-time }
      required: [orgId, userId, role]
...
    OrgInvite:
      type: object
      required: [email, role]
      properties:
        email: { type: string, format: email }
        role:
          type: string
          enum: [OWNER, ADMIN, MEMBER, VIEWER]
    OrgMemberPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OrgMember' }
          required: [items]

    # ---- rbac ----
...
    LicenseToken:
      type: object
      properties:
        jti: { type: string }
        subscriptionId: { type: string, format: uuid }
        orgId: { type: string, format: uuid }
        kid: { type: string }
        plan: { type: string }
        permissions:
          type: array
          items: { type: string }
        features:
          type: object
          additionalProperties: true
        seats: { type: integer }
        issuedAt: { type: string, format: date-time }
        notBefore: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokeReason: { type: string, nullable: true }
        fingerprint: { type: string, description: 'SHA-256 of the compact JWT.' }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        lastSeenIp: { type: string, nullable: true }
        status:
          type: string
          enum: [ACTIVE, REVOKED, EXPIRED]
      required: [jti, subscriptionId, kid, issuedAt, expiresAt, status]
    LicenseIssueRequest:
      type: object
      properties:
        audience:
          type: string
          description: Target audience claim. Defaults to the subscription's configured audience.
        ttlDays:
          type: integer
          description: Override the plan's default TTL.
        notBefore: { type: string, format: date-time, nullable: true }
        notes: { type: string }
        kid:
          type: string
          description: Force signing with a specific key. Defaults to the active key.
    LicenseIssueResponse:
      type: object
      required: [token, envelope]
      properties:
        token: { $ref: '#/components/schemas/LicenseToken' }
        envelope: { $ref: '#/components/schemas/LicenseEnvelope' }
        downloadFilename: { type: string, example: 'acme-corp-pro-2026-06-01.lic' }
    LicenseEnvelope:
      type: object
      required: [license]
      properties:
        license: { type: string, description: 'The compact-serialized signed JWT.' }
        issued_at: { type: string, format: date-time }
        customer: { type: string }
        plan: { type: string }
        expires_at: { type: string, format: date-time }
        notes: { type: string }
...